+++ examples/petstore/settings.py
@@ -1,0 +1,8 @@
+import attr
+
+from rororo.settings import BaseSettings, env_factory
+
+
+@attr.dataclass(frozen=True, slots=True)
+class Settings(BaseSettings):
+    pets_app_key: str = env_factory("PETS_APP_KEY", "pets")