+++ tests/test_settings.py
@@ -2,11 +2,16 @@ import calendar
 import datetime
 import logging
 import os
+from pathlib import Path
+from typing import Optional
 
+import attr
 import pytest
 
 from rororo.logger import default_logging_dict
 from rororo.settings import (
+    BaseSettings,
+    env_factory,
     from_env,
     immutable_settings,
     inject_settings,
@@ -37,6 +42,104 @@ def check_immutability(settings):
     # Cannot update values at all
     with pytest.raises(AttributeError):
         settings.update({key: "new-value", "TEST_SETTING": "test_value"})
+
+
+def test_base_settings():
+    settings = BaseSettings()
+    assert settings.host == "localhost"
+    assert settings.port == 8080
+    assert settings.debug is False
+    assert settings.level == "test"
+    assert settings.time_zone == "UTC"
+    assert settings.first_weekday == 0
+    assert settings.locale == "en_US.UTF-8"
+    assert settings.sentry_dsn is None
+    assert settings.sentry_release is None
+
+
+def test_base_settings_from_env(monkeypatch):
+    monkeypatch.setenv("DEBUG", "on")
+    assert BaseSettings().debug is True
+
+
+def test_base_settings_from_kwargs():
+    assert BaseSettings(debug=True).debug is True
+
+
+def test_base_settings_inheritance(monkeypatch):
+    monkeypatch.setenv("USE_RORORO", "yes")
+
+    @attr.dataclass(frozen=True, slots=True)
+    class Settings(BaseSettings):
+        use_rororo: bool = env_factory("USE_RORORO", False)
+
+    settings = Settings()
+    assert settings.debug is False
+    assert settings.use_rororo is True
+
+
+@pytest.mark.parametrize(
+    "name, expected", (("LEVEL", "test"), ("dOesNotExist", None))
+)
+def test_env_factory(name, expected):
+    @attr.dataclass
+    class Model:
+        value: Optional[str] = env_factory(name)
+
+    instance = Model()
+    assert instance.value == expected
+
+
+@pytest.mark.parametrize(
+    "str_value, expected",
+    (("yes", True), ("off", False), ("no", False), ("0", False), ("1", True)),
+)
+def test_env_factory_bool(monkeypatch, str_value, expected):
+    monkeypatch.setenv("USE_SOMETHING", str_value)
+
+    @attr.dataclass
+    class Model:
+        use_something: bool = env_factory("USE_SOMETHING", False)
+
+    instance = Model()
+    assert instance.use_something is expected
+
+
+def test_env_factory_custom_type(monkeypatch):
+    monkeypatch.setenv("DATA_PATH", "/tmp/data")
+
+    @attr.dataclass
+    class Model:
+        data_path: Path = env_factory(
+            "DATA_PATH", Path("~").expanduser() / "data"
+        )
+
+    instance = Model()
+    assert instance.data_path == Path("/") / "tmp" / "data"
+
+
+@pytest.mark.parametrize(
+    "name, default, expected",
+    (("LEVEL", "dev", "test"), ("dOesNotExist", "ok", "ok")),
+)
+def test_env_factory_default(name, default, expected):
+    @attr.dataclass
+    class Model:
+        value: str = env_factory(name, default)
+
+    instance = Model()
+    assert instance.value == expected
+
+
+def test_env_factory_invalid_type(monkeypatch):
+    monkeypatch.setenv("FIRST_WEEKDAY", "one")
+
+    @attr.dataclass
+    class Model:
+        value: str = env_factory("FIRST_WEEKDAY", 0)
+
+    with pytest.raises(ValueError):
+        Model()
 
 
 def test_from_env():